Add mass conservation and bubble cross model
We include additional modules for vesicle evolution using a quadratic expression with distance and a bubble rheology model that implements a capillary number-dependent rheology of the Cross-type model. We make changes to the base code to iteratively calcualte viscosity and velocity. Additionally, we change the flow width calcualtion to conserve mass rather than volume.